Rename Components to JSX
